# 8326962: C2 SuperWord: cache VPointer