--- a/src/hotspot/share/opto/superword.hpp
+++ b/src/hotspot/share/opto/superword.hpp
@@ -495,6 +495,11 @@ class SuperWord : public ResourceObj {
     return _vloop_analyzer.dependency_graph().mutually_independent(nodes);
   }
 
+  // VLoopDependencyGraph Accessors
+  const VPointer& get_pointer(const MemNode* mem) const {
+    return _vloop_analyzer.pointers().get(mem);
+  }
+
 #ifndef PRODUCT
   // TraceAutoVectorization and TraceSuperWord
   bool is_trace_superword_alignment() const {