--- a/src/hotspot/share/opto/traceAutoVectorizationTag.hpp
+++ b/src/hotspot/share/opto/traceAutoVectorizationTag.hpp
@@ -29,12 +29,13 @@
 #include "utilities/stringUtils.hpp"
 
 #define COMPILER_TRACE_AUTO_VECTORIZATION_TAG(flags) \
-  flags(POINTER_ANALYSIS,     "Trace VPointer") \
+  flags(POINTER_ANALYSIS,     "Trace VPointer (verbose)") \
   flags(PRECONDITIONS,        "Trace VLoop::check_preconditions") \
   flags(LOOP_ANALYZER,        "Trace VLoopAnalyzer::setup_submodules") \
   flags(MEMORY_SLICES,        "Trace VLoopMemorySlices") \
   flags(BODY,                 "Trace VLoopBody") \
   flags(TYPES,                "Trace VLoopTypes") \
+  flags(POINTERS,             "Trace VLoopPointers") \
   flags(DEPENDENCY_GRAPH,     "Trace VLoopDependencyGraph") \
   flags(SW_ALIGNMENT,         "Trace SuperWord alignment analysis") \
   flags(SW_ADJACENT_MEMOPS,   "Trace SuperWord::find_adjacent_refs") \