This is a subtask of [JDK-8315361](https://bugs.openjdk.org/browse/JDK-8315361).

Parsing `VPointer` currently happens all over SuperWord. And often in quadratic loops, where we compare all-with-all loads/stores.

I propose to cache the `VPointer`s, then we can do a constant-time cache lookup rather than parsing the pointer subgraph every time.

There are now only a few cases where we cannot use the cached `VPointer`:
- `SuperWord::unrolling_analysis`: we have no `VLoopAnalyzer`, and so no submodules like `VLoopPointers`. We don't need to cache, since we only iterate over the loop body once, and create only a single `VPointer` per memop.
- `SuperWord::output`: when we have a `Load`, and try to bypass `StoreVector` nodes. The `StoreVector` nodes are new, and so we have no cached `VPointer` for them. This could be fixed somehow, but I don't want to deal with it now. I intend to refactor `SuperWord::output` soon, and can look into options at that point (either I bypass before we insert the vector nodes, or I remember what scalar memop the vector was created from, and then get the cached pointer this way).

This changeset is also a preparation step for [JDK-8325155](https://bugs.openjdk.org/browse/JDK-8325155). I will have a list of pointers, and sort them such that creating adjacent refs is much more efficient.

**Benchmarking SuperWord Compile Time**

I use the same benchmark from https://github.com/openjdk/jdk/pull/18532.

On master:
```
    C2 Compile Time:       56.816 s
         IdealLoop:            56.604 s
           AutoVectorize:      56.192 s
```

With this patch:
```
    C2 Compile Time:       49.719 s
         IdealLoop:            49.509 s
           AutoVectorize:      49.106 s
```

This saves us about `7 sec`, which is significant. I will have to see what it effect it has once we also apply https://github.com/openjdk/jdk/pull/18532, but I think the combined effect will be very significant.

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8326962](https://bugs.openjdk.org/browse/JDK-8326962): C2 SuperWord: cache VPointer (**Sub-task** - P4)


### Reviewers
 * [Christian Hagedorn](https://openjdk.org/census#chagedorn) (@chhagedorn - **Reviewer**)
 * [Vladimir Kozlov](https://openjdk.org/census#kvn) (@vnkozlov - **Reviewer**)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk.git pull/18577/head:pull/18577` \
`$ git checkout pull/18577`

Update a local copy of the PR: \
`$ git checkout pull/18577` \
`$ git pull https://git.openjdk.org/jdk.git pull/18577/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 18577`

View PR using the GUI difftool: \
`$ git pr show -t 18577`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/18577.diff">https://git.openjdk.org/jdk/pull/18577.diff</a>

</details>


### Webrev
[Link to Webrev Comment](https://git.openjdk.org/jdk/pull/18577#issuecomment-2032002927)