@eme64 2024-04-03
> One question: will VLoopAnalyzer default destructor clean up all memory used?

@vnkozlov there is no need, since it is all allocated over the `Arena` in `VLoopAnalyzer`:

```
// Arena for all submodules
Arena                _arena;
```

It is that arena that I pass into all submodules, such as `VLoopVPointer`. `VLoopAnalyzer` is stack allocated, so once the destructor removes its `_arena`, all submodules are also automatically deallocated.