@vnkozlov 2024-04-03
> It is that arena that I pass into all submodules, such as `VLoopVPointer`. `VLoopAnalyzer` is stack allocated, so once the destructor removes its `_arena`, all submodules are also automatically deallocated.

Good.