@eme64 2024-04-04
Thanks @vnkozlov @chhagedorn @jdksjolen for the reviews and suggestions!
@jdksjolen feel free to give me your ideas about Arena-allocation, I can still improve in a follow-up RFE ;)

/integrate